Treating exceptions

+++ src/main/java/com/yoaceng/yocommerce/services/ProductService.java
@@ -3,11 +3,16 @@ package com.yoaceng.yocommerce.services;
 import com.yoaceng.yocommerce.dto.ProductDTO;
 import com.yoaceng.yocommerce.entities.Product;
 import com.yoaceng.yocommerce.repositories.ProductRepository;
+import com.yoaceng.yocommerce.services.exceptions.DatabaseException;
+import com.yoaceng.yocommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 
@@ -21,7 +26,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Resource not found by the given id."));
         return convertToDto(product);
     }
 
@@ -41,16 +47,27 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product productById = repository.getReferenceById(id);
-        // Need to use this different method because we need to keep the original entity
-        copyDtoToEntity(dto, productById);
-        Product result = repository.save(productById);
-        return convertToDto(result);
+        try{
+            Product productById = repository.getReferenceById(id);
+            copyDtoToEntity(dto, productById);
+            Product result = repository.save(productById);
+            return convertToDto(result);
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
-    public void delete (long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity){
